pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wizerd"
version = "0.1.0"
description = "PostgreSQL ER diagram generator with zero-overlap layouts"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
keywords = [
    "database",
    "postgresql",
    "erd",
    "diagram",
    "schema",
    "visualization",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Documentation",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.1.7",
    "typer>=0.12.3",
    "sqlparse>=0.5.0",
    "svgwrite>=1.4.3",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
export = ["cairosvg>=2.7.1"]
dev = [
    "pytest>=8.3.2",
    "ruff>=0.5.4",
    "mypy>=1.10.0",
    "build>=1.0.0",
    "twine>=4.0.0",
    "pyinstaller>=6.0.0",
    "python-semantic-release>=9.0.0",
]

[tool.setuptools.package-data]
"wizerd" = ["py.typed"]
"wizerd.templates.config" = ["*.yaml"]
"wizerd.theme" = ["presets.json"]
"wizerd.layout" = ["*.mjs", "package.json", "node_modules/**/*"]

[tool.setuptools.packages.find]
where = ["."]
include = ["wizerd*"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["E203", "E501"]

[tool.mypy]
python_version = "3.9"
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
ignore_missing_imports = true
no_implicit_optional = false
